pyproject.toml
[project]
name = "finance-research-agent"
version = "0.1.0"
description = "Financial research agent optimized for the Vals AI Finance Agent Benchmark"
requires-python = ">=3.12"

dependencies = [
    "anthropic>=0.52.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "datasets>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
